+++ trees-mutable-binary/src/test/java/com/mattunderscore/trees/binary/mutable/MutableBinaryTreeImplTest.java
@@ -23,7 +23,7 @@ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 
-package com.mattunderscore.trees.binary;
+package com.mattunderscore.trees.binary.mutable;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -34,6 +34,7 @@ import java.util.Iterator;
 
 import org.junit.Test;
 
+import com.mattunderscore.trees.binary.MutableBinaryTreeNode;
 import com.mattunderscore.trees.binary.mutable.MutableBinaryTreeImpl;
 
 /**