+++ trees-mutable-binary/src/test/java/com/mattunderscore/trees/binary/mutable/MutableBinaryTreeIteratorTest.java
@@ -23,7 +23,7 @@ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 
-package com.mattunderscore.trees.binary;
+package com.mattunderscore.trees.binary.mutable;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -33,6 +33,7 @@ import java.util.Iterator;
 import org.junit.Test;
 
 import com.mattunderscore.trees.Trees;
+import com.mattunderscore.trees.binary.MutableBinaryTreeNode;
 import com.mattunderscore.trees.binary.mutable.MutableBinaryTreeImpl;
 import com.mattunderscore.trees.impl.TreesImpl;
 import com.mattunderscore.trees.tree.Tree;
@@ -40,7 +41,7 @@ import com.mattunderscore.trees.tree.Tree;
 /**
  * @author Matt Champion on 08/09/14.
  */
-public final class BinaryTreeIteratorTest {
+public final class MutableBinaryTreeIteratorTest {
     private static final Trees trees = Trees.get();
 
     @Test