+++ CHANGELOG.md
@@ -117,3 +117,4 @@ source nodes
 * Renamed iterators package
 * Changed the return type of the TreeNodeImpl child iterator as it is final
 * Added SPI component to allow tree implementations to be discovered 
+* Moved mutable binary trees into separate module